requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
python-frontmatter>=1.1.0
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=0.25.0
numpy>=1.26.0
